Show spots on the calendar
