+++ Caca/CalendarCellView.xib
@@ -18,7 +18,7 @@
             <subviews>
                 <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="8xL-ki-MaE">
                     <rect key="frame" x="18.5" y="28.5" width="163.5" height="163.5"/>
-                    <color key="backgroundColor" white="0.66666666666666663" alpha="1" colorSpace="calibratedWhite"/>
+                    <color key="backgroundColor" red="0.8862745098" green="0.87843137250000003" blue="0.64705882349999999" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                     <constraints>
                         <constraint firstAttribute="width" secondItem="8xL-ki-MaE" secondAttribute="height" multiplier="1:1" id="uAw-el-gfF"/>
                     </constraints>
