+++ Caca/FillinViewController.swift
@@ -217,10 +217,12 @@ class FillinViewController: UIViewController {
         let month = calendar.component(.month, from: date)
         let day = calendar.component(.day, from: date)
 
-        let hour = calendar.component(.hour, from: date)
-        let minute = calendar.component(.minute, from: date)
+//        let hour = calendar.component(.hour, from: date)
+//        let minute = calendar.component(.minute, from: date)
 
-        return String(format: "%04i-%02i-%02i %02i:%02i", year, month, day, hour, minute)
+//        return String(format: "%04i-%02i-%02i %02i:%02i", year, month, day, hour, minute)
+        return String(format: "%04i-%02i-%02i", year, month, day)
+
     }
 
     func hideKeyBoard() {
