+++ Caca/Palette.swift
@@ -14,4 +14,6 @@ struct Palette {
 
     static let textColor = UIColor(red: 131/255, green: 91/255, blue: 0/255, alpha: 1)
 
+    static let todayColor = UIColor(red: 226/255, green: 224/255, blue: 165/255, alpha: 1)
+
 }
