+++ Caca/RecordTableViewController.swift
@@ -26,7 +26,18 @@ class RecordTableViewController: UITableViewController {
 
         tableView.register(RecordTableViewCell.self, forCellReuseIdentifier: "RecordTableViewCell")
 
-        getCaca()
+        let provider = CacaProvider.shared
+
+        provider.getCaca { (cacas, _) in
+
+            if let cacas = cacas {
+                self.cacas = cacas
+            }
+            self.tableView.reloadData()
+
+        }
+
+//        getCaca()
     }
 
     // MARK: Table view data source
@@ -81,39 +92,39 @@ class RecordTableViewController: UITableViewController {
         self.navigationController?.pushViewController(recordDetailViewController, animated: true)
     }
 
-    // MARK: Get cacaInfo from Firebase
-    func getCaca() {
-
-        let rootRef = FIRDatabase.database().reference()
-
-        rootRef.child("cacas").queryOrdered(byChild: "host").queryEqual(toValue: FIRAuth.auth()?.currentUser?.uid).observeSingleEvent(of: .value, with: { (snapshot) in
-            if let snaps = snapshot.children.allObjects as? [FIRDataSnapshot] {
-
-                for snap in snaps {
-
-                    if let cacaInfo = snap.value as? NSDictionary,
-                        let cacaPhoto = cacaInfo["photo"] as? String,
-                        let cacaDate = cacaInfo["date"] as? String,
-                        let cacaTime = cacaInfo["consumingTime"] as? String,
-                        let cacaShape = cacaInfo["shape"] as? Int,
-                        let cacaColor = cacaInfo["color"] as? Int,
-                        let cacaAmount = cacaInfo["amount"] as? Double,
-                        let cacaOther = cacaInfo["other"] as? String {
-
-                        let caca = Caca(photo: cacaPhoto, date: cacaDate, consumingTime: cacaTime, shape: Shape(rawValue: cacaShape)!, color: Color(rawValue: cacaColor)!, amount: cacaAmount, otherInfo: cacaOther)
-
-                        self.cacas.append(caca)
-                    }
-                }
-            }
-
-            DispatchQueue.main.async {
-                self.tableView.reloadData()
-            }
-
-        }) { (error) in
-            print(error.localizedDescription)
-        }
-    }
+//    // MARK: Get cacaInfo from Firebase
+//    func getCaca() {
+//
+//        let rootRef = FIRDatabase.database().reference()
+//
+//        rootRef.child("cacas").queryOrdered(byChild: "host").queryEqual(toValue: FIRAuth.auth()?.currentUser?.uid).observeSingleEvent(of: .value, with: { (snapshot) in
+//            if let snaps = snapshot.children.allObjects as? [FIRDataSnapshot] {
+//
+//                for snap in snaps {
+//
+//                    if let cacaInfo = snap.value as? NSDictionary,
+//                        let cacaPhoto = cacaInfo["photo"] as? String,
+//                        let cacaDate = cacaInfo["date"] as? String,
+//                        let cacaTime = cacaInfo["consumingTime"] as? String,
+//                        let cacaShape = cacaInfo["shape"] as? Int,
+//                        let cacaColor = cacaInfo["color"] as? Int,
+//                        let cacaAmount = cacaInfo["amount"] as? Double,
+//                        let cacaOther = cacaInfo["other"] as? String {
+//
+//                        let caca = Caca(photo: cacaPhoto, date: cacaDate, consumingTime: cacaTime, shape: Shape(rawValue: cacaShape)!, color: Color(rawValue: cacaColor)!, amount: cacaAmount, otherInfo: cacaOther)
+//
+//                        self.cacas.append(caca)
+//                    }
+//                }
+//            }
+//
+//            DispatchQueue.main.async {
+//                self.tableView.reloadData()
+//            }
+//
+//        }) { (error) in
+//            print(error.localizedDescription)
+//        }
+//    }
 
 }
